// lib/pokeapi-types.ts
        interface NamedAPIResource {
            /** The name of the referenced resource. */
            name: string;
            /** The URL of the referenced resource. */
            url: string;
            [property: string]: any;
        }
// ...
        interface VersionGameIndex {
            /** The internal id of an API resource within game data. */
            game_index: number;
            /** The version relevent to this game index. */
            version: NamedAPIResource;
            [property: string]: any;
        }
// ...
        interface RarityVersion {
          rarity: number;
          version: NamedAPIResource;
          [property: string]: any;
      }
// ...
        interface Pokemon {
            /** A list of abilities this Pokémon could potentially have. */
            abilities: PokemonAbility[];
            /** The base experience gained for defeating this Pokémon. */
            base_experience: number;
            /** A set of cries used to depict this Pokémon in the game. A visual representation of the various cries can be found at <a href='https://github.com/PokeAPI/cries#cries'>PokeAPI/cries</a> */
            cries: Cries;
            /** A list of forms this Pokémon can take on. */
            forms: NamedAPIResource[];
            /** A list of game indices relevent to Pokémon item by generation. */
            game_indices: VersionGameIndex[];
            /** The height of this Pokémon in decimetres. */
            height: number;
            /** A list of items this Pokémon may be holding when encountered. */
            held_items: HeldItem[];
            /** The identifier for this resource. */
            id: number;
            /** Set for exactly one Pokémon used as the default for each species. */
            is_default: boolean;
            /** A link to a list of location areas, as well as encounter details pertaining to specific versions. */
            location_area_encounters: string;
            /** A list of moves along with learn methods and level details pertaining to specific version groups. */
            moves: MoveElement[];
            /** The name for this resource. */
            name: string;
            /** Order for sorting. Almost national order, except families are grouped together. */
            order: number;
            past_abilities: PastAbility[];
            /** A list of details showing types this pokémon had in previous generations */
            past_types: PastType[];
            /** The species this Pokémon belongs to. */
            species: NamedAPIResource;
            /** A set of sprites used to depict this Pokémon in the game. A visual representation of the various sprites can be found at <a href='https://github.com/PokeAPI/sprites#sprites'>PokeAPI/sprites</a> */
            sprites: PokemonSprites;
            /** A list of base stat values for this Pokémon. */
            stats: StatElement[];
            /** A list of details showing types this Pokémon has. */
            types: PokemonType[];
            /** The weight of this Pokémon in hectograms. */
            weight: number;
            [property: string]: any;
        }

        interface PokemonAbility {
            /** The ability the Pokémon may have. */
            ability: NamedAPIResource;
            /** Whether or not this is a hidden ability. */
            is_hidden: boolean;
            /** The slot this ability occupies in this Pokémon species. */
            slot: number;
            [property: string]: any;
        }

        interface Cries {
            latest: string;
            legacy: null | string;
            [property: string]: any;
        }

        interface HeldItem {
            item: NamedAPIResource;
            version_details: RarityVersion[];
            [property: string]: any;
        }

        interface MoveElement {
            move: NamedAPIResource;
            version_group_details: VersionGroupDetail[];
            [property: string]: any;
        }

        interface VersionGroupDetail {
            level_learned_at: number;
            move_learn_method: NamedAPIResource;
            version_group: NamedAPIResource;
            [property: string]: any;
        }

        interface PastAbility {
            abilities: PastAbilityAbility[];
            generation: NamedAPIResource;
            [property: string]: any;
        }

        interface PastAbilityAbility {
            ability: NamedAPIResource;
            is_hidden: boolean;
            slot: number;
            [property: string]: any;
        }

        interface PastType {
            generation: NamedAPIResource;
            types: PastTypeType[];
            [property: string]: any;
        }

        interface PastTypeType {
            slot: number;
            type: NamedAPIResource;
            [property: string]: any;
        }

        interface PokemonSprites {
            /** The default depiction of this Pokémon from the back in battle. */
            back_default: null | string;
            /** The female depiction of this Pokémon from the back in battle. */
            back_female: null | string;
            /** The shiny depiction of this Pokémon from the back in battle. */
            back_shiny: null | string;
            /** The shiny female depiction of this Pokémon from the back in battle. */
            back_shiny_female: null | string;
            /** The default depiction of this Pokémon from the front in battle. */
            front_default: null | string;
            /** The female depiction of this Pokémon from the front in battle. */
            front_female: null | string;
            /** The shiny depiction of this Pokémon from the front in battle. */
            front_shiny: null | string;
            /** The shiny female depiction of this Pokémon from the front in battle. */
            front_shiny_female: null | string;
            other: Other;
            versions: SpriteVersions;
            [property: string]: any;
        }

        interface Other {
            dream_world: DreamWorld;
            home: Home;
            "official-artwork": OfficialArtwork;
            showdown: Showdown;
            [property: string]: any;
        }
// ...
        interface SpriteVersions {
            "generation-i": GenerationI;
            "generation-ii": GenerationIi;
            "generation-iii": GenerationIii;
            "generation-iv": GenerationIv;
            "generation-v": GenerationV;
            "generation-vi": GenerationVi;
            "generation-vii": GenerationVii;
            "generation-viii": GenerationViii;
            [property: string]: any;
        }
// ...
        interface StatElement {
            base_stat: number;
            effort: number;
            stat: NamedAPIResource;
            [property: string]: any;
        }

        interface PokemonType {
            /** The order the Pokémon's types are listed in. */
            slot: number;
            /** The type the referenced Pokémon has. */
            type: NamedAPIResource;
            [property: string]: any;
        }